+++ core/template/pug4j/src/test/kotlin/org/http4k/template/Pug4jTemplatesTest.kt
@@ -3,6 +3,7 @@ package org.http4k.template
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 
 class Pug4JTemplatesTest : TemplatesContract<Pug4jTemplates>(Pug4jTemplates()) {
     @Test
@@ -16,6 +17,15 @@ class Pug4JTemplatesTest : TemplatesContract<Pug4jTemplates>(Pug4jTemplates()) {
     fun `multiple loaders, first loaded wins`() {
         val rendererBFirst = templates.HotReload("src/test/resources/b")
         assertThat(rendererBFirst(TemplateC), equalTo("c2"))
+    }
+
+    @Test
+    fun `hot reload rejects template path that escapes base dir`() {
+        val renderer = templates.HotReload("src/test/resources/a")
+        val escaping = object : ViewModel {
+            override fun template() = "../AtRootBob"
+        }
+        assertThrows<ViewNotFound> { renderer(escaping) }
     }
 }